+++ libfsntfs/libfsntfs_directory_entry.h
@@ -64,7 +64,7 @@ int libfsntfs_directory_entry_compare(
 
 int libfsntfs_directory_entry_get_mft_entry_index(
      libfsntfs_directory_entry_t *directory_entry,
-     int *mft_entry_index,
+     uint64_t *mft_entry_index,
      libcerror_error_t **error );
 
 #if defined( __cplusplus )
